+++ frontend/app/page.tsx
@@ -1,4 +1,4 @@
-import { Header } from "./components/elements/header/header";
+import { Header } from "./components/elements/Header/Header";
 import { ItemsCard } from "./components/elements/ItemsCard/ItemsCard";
 
 export default function Home() {
